+++ include/compile.h
@@ -5,6 +5,6 @@
 #define _inline inline __attribute__((always_inline))
 
 #define containof(ptr, type, member)\
-	((type *)((char *)(ptr) - (int)&((type *)0)->member))
+	((type *)((char *)(ptr) - (size_t)&((type *)0)->member))
 
 #endif	/* complie.h */
